feat:fix dashboard page , change profile name api, add style to dashboard page
