+++ src/pages/dashboard/my-coupons/index.js
@@ -1,0 +1,75 @@
+import React, { useEffect, useState } from "react";
+import { useRouter } from "next/router";
+import axios from "axios";
+import Navbar from "@/components/Navbar/Navbar";
+import { Button } from "@/components/ui/button";
+import CouponsCard from "@/components/Coupons-components/CouponsCard";
+
+const MyCoupon = () => {
+  const router = useRouter();
+  const [couponsInWallet, setCouponsInWallet] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  console.log(couponsInWallet);
+
+  const fetchCoupons = async () => {
+    try {
+      const response = await axios.get(
+        "/api/dashboard/get-all-from-user-coupon"
+      );
+      setCouponsInWallet(response.data.data);
+      setLoading(false);
+    } catch (error) {
+      setError(error.response?.data?.error || error.message);
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchCoupons();
+  }, []);
+
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center min-h-screen">
+        <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-gray-900"></div>
+      </div>
+    );
+  }
+
+  if (error) {
+    return (
+      <div className="text-center text-red-500 p-4">
+        <p>เกิดข้อผิดพลาด: {error}</p>
+      </div>
+    );
+  }
+
+  return (
+    <>
+      <div className="container mx-auto px-4 py-8">
+        <h1 className="text-2xl font-bold mb-6">My coupons</h1>
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
+          {couponsInWallet.map((coupon) => (
+            <div key={coupon.coupon_id} className="relative">
+              <CouponsCard
+                coupon_id={coupon.coupon_id}
+                image={coupon.coupons.image}
+                title={coupon.coupons.title}
+                end_date={coupon.coupons.end_date}
+              />
+            </div>
+          ))}
+        </div>
+        {couponsInWallet.length === 0 && (
+          <div className="text-center text-gray-500 mt-8">
+            <p>คุณยังไม่มีคูปองในกระเป๋า</p>
+          </div>
+        )}
+      </div>
+    </>
+  );
+};
+
+export default MyCoupon;
